Monthly Salary Report

+++ HRMS_Core/Salary/SalaryDetailViewModel.cs
@@ -40,7 +40,7 @@ namespace HRMS_Core.Salary
         public decimal? AbsentDays { get; set; }
         public decimal? PresentDays { get; set; }
         public int? WeekOff { get; set; }
-        public int? Leave { get; set; }
+        public decimal? Leave { get; set; }
         public int? SalaryDays { get; set; }
         public string? BranchName { get; set; }
     }
